Add method deleteRoot and supporting functions
Created deleteRoot method to delete the specified root.
Created the getPreviousRoot function to get the previous root of the
node that will be deleted. The said function helps on the method
deleteRoot.
Created the getAppropriateLeaf function to get the suitable root node
to replace the to be deleted root. This function also healps on the
method deleteRoot.