Add levelOrder method
Created treeData that will hold the values of the data structure.
Created queueArray that act as a data storage for processing the
values of the data structure through first in first out method.
Created levelOrder method that will take in a function and provide
the data in levelOrder as an argument to the function. If the
function is not provided, the method will return an array (treeData)
containing a level order values of the data structure.
Created getLevelOrderData function that will process the data
structure and get the values in a levelOrder sequence.