Add Depth-first traversal and height methods.
Created inorder, preorder, postorder methods that will take in a
function and provide the data in appropriate sequence as an argument
to the function. If the function is not provided, the method will
return an array (treeData).
Created the height method to get the height from the node to the
farthest leaf node which is the data structure (level -1).
Added level variable to store the count how many levels a node has.
Added levelArray to store all the nodes per level.
Added countLevel function that will count all the level in the data
structure.